Refactor the test suite, add performance spotcheck tests for profiling.
