+++ pom.xml
@@ -39,10 +39,7 @@
       </plugin>
       <plugin>
         <artifactId>maven-surefire-plugin</artifactId>
-        <version>2.20</version>
-        <configuration>
-          <forkMode>never</forkMode>
-        </configuration>
+        <version>2.19.1</version>
       </plugin>
       <plugin>
         <groupId>org.codehaus.mojo</groupId>
